Fix hotkeys dialog registration to prevent null reference error
- Move registerHelpModal() from project-tab component back to app.component
- Prevents race condition where component initializes before project input is set
- Fix dialog close button to use mat-flat-button with primary color
- Matches styling of other dialogs in the application

Fixes issue introduced in commit 2061e4c where moving the hotkeys registration
to the project-tab component caused 'Cannot read properties of null' errors.
