+++ src/app/shared/components/hotkeys-help/hotkeys-help-dialog.component.html
@@ -31,5 +31,5 @@ Copyright 2021 Carnegie Mellon University. All Rights Reserved.
 </mat-dialog-content>
 
 <mat-dialog-actions align="end">
-  <button mat-button (click)="close()">Close</button>
+  <button mat-flat-button color="primary" (click)="close()">Close</button>
 </mat-dialog-actions>
